move all providers to containers dir
ubergrape/chatgrape#4131
