+++ src/app/index.js
@@ -11,19 +11,20 @@ import wrapWithIntlProvider from './wrapWithIntlProvider'
 import * as translations from '../i18n'
 import subscribe from './subscribe'
 import {create as createClient} from '../utils/backend/client'
-import BillingWarningProvider from '../components/billing-warning/BillingWarningProvider'
-import TypingNotificationProvider from '../components/typing-notification/TypingNotificationProvider'
-import AlertsProvider from '../components/alerts/AlertsProvider'
-import ChannelMembersInviteProvider from '../components/channel-members-invite/ChannelMembersInviteProvider'
-import NewConversationProvider from '../components/new-conversation/NewConversationProvider'
-import UnreadChannelsProvider from '../components/unread-channels/UnreadChannelsProvider'
-import OrgInfoProvider from '../components/org-info/OrgInfoProvider'
-import NavigationProvider from '../components/navigation/NavigationProvider'
-import HeaderProvider from '../components/header/HeaderProvider'
-import SidebarProvider from '../components/redux-sidebar/SidebarProvider'
-import HistoryProvider from '../components/redux-history/HistoryProvider'
-import InviteToOrgProvider from '../components/redux-invite-to-org/InviteToOrgProvider'
-import NotificationSettingsProvider from '../components/redux-notification-settings/NotificationSettingsProvider'
+
+import {BillingWarningProvider} from '../containers/billing-warning'
+import {TypingNotificationProvider} from '../containers/typing-notification'
+import {AlertsProvider} from '../containers/alerts'
+import {ChannelMembersInviteProvider} from '../containers/channel-members-invite'
+import {NewConversationProvider} from '../containers/new-conversation'
+import {UnreadChannelsProvider} from '../containers/unread-channels'
+import {OrgInfoProvider} from '../containers/org-info'
+import {NavigationProvider} from '../containers/navigation'
+import {HeaderProvider} from '../containers/header'
+import {SidebarProvider} from '../containers/sidebar'
+import {HistoryProvider} from '../containers/history'
+import {InviteToOrgProvider} from '../containers/invite-to-org'
+import {NotificationSettingsProvider} from '../containers/notification-settings'
 
 const {languageCode, email, username, id: userId} = user
 const messages = translations[languageCode]
