--- src/components/alerts/actionNames.js
@@ -1,5 +1,0 @@
-export default [
-  'enableNotifications',
-  'hideAlert',
-  'clearAlertDelay'
-]
