--- src/components/billing-warning/actionNames.js
@@ -1,5 +1,0 @@
-export default [
-  'showBillingWarning',
-  'hideBillingWarning',
-  'goToPayment'
-]
