--- src/components/channel-members-invite/actionNames.js
@@ -1,9 +1,0 @@
-export default [
-  'showInviteToOrg',
-  'inviteToChannel',
-  'createRoomFromPmAndInvite',
-  'hideChannelMembersInvite',
-  'addToChannelMembersInvite',
-  'removeFromChannelMembersInvite',
-  'setInviteFilterValue'
-]
