--- src/components/header/actionNames.js
@@ -1,9 +1,0 @@
-export default [
-  'showChannelMembersInvite',
-  'showInSidebar',
-  'hideSidebar',
-  'requestAddChannelToFavorites',
-  'requestRemoveChannelFromFavorites',
-  'updateMessageSearchQuery',
-  'hideIntercom'
-]
