--- src/components/navigation/actionNames.js
@@ -1,7 +1,0 @@
-export default [
-  'showChannelsManager',
-  'showPmManager',
-  'showNewConversation',
-  'goToChannel',
-  'focusGrapeInput'
-]
