--- src/components/new-conversation/actionNames.js
@@ -1,12 +1,0 @@
-export default [
-  'goToChannel',
-  'createRoomWithUsers',
-  'showInviteToOrg',
-  'showNewConversation',
-  'hideNewConversation',
-  'showNewConversationAdvanced',
-  'addToNewConversation',
-  'removeFromNewConversation',
-  'filterNewConversation',
-  'clearRoomCreateError'
-]
