--- src/components/org-info/actionNames.js
@@ -1,3 +1,0 @@
-export default [
-  'toggleOrgSettings'
-]
