--- src/components/typing-notification/actionNames.js
@@ -1,4 +1,0 @@
-export default [
-  'setUsers',
-  'cleanupTyping'
-]
