+++ src/containers/alerts/AlertsProvider.js
@@ -2,10 +2,15 @@ import React, {Component} from 'react'
 import {Provider, connect} from 'react-redux'
 
 import {mapActionsToProps} from '../../app/redux'
-import {alertsAndChannelSelector as selector} from '../../selectors'
 import getStore from '../../app/store'
-import actionNames from './actionNames'
-import Alerts from './Alerts'
+import {alertsAndChannelSelector as selector} from '../../selectors'
+import Alerts from '../../components/alerts/Alerts'
+
+const actionNames = [
+  'enableNotifications',
+  'hideAlert',
+  'clearAlertDelay'
+]
 
 const ConnectedAlerts = connect(
   selector,
