+++ src/containers/alerts/index.js
@@ -1,0 +1,1 @@
+export AlertsProvider from './AlertsProvider'
