+++ src/containers/billing-warning/BillingWarningProvider.js
@@ -4,8 +4,13 @@ import {Provider, connect} from 'react-redux'
 import {mapActionsToProps} from '../../app/redux'
 import {billingWarningSelector} from '../../selectors'
 import getStore from '../../app/store'
-import actionNames from './actionNames'
-import BillingWarning from './BillingWarning'
+import BillingWarning from '../../components/billing-warning/BillingWarning'
+
+const actionNames = [
+  'showBillingWarning',
+  'hideBillingWarning',
+  'goToPayment'
+]
 
 const ConnectedBillingWarning = connect(
   billingWarningSelector,
