+++ src/containers/channel-members-invite/index.js
@@ -1,0 +1,1 @@
+export ChannelMembersInviteProvider from './ChannelMembersInviteProvider'
