+++ src/containers/header/HeaderProvider.js
@@ -2,10 +2,19 @@ import React, {Component} from 'react'
 import {Provider, connect} from 'react-redux'
 
 import {mapActionsToProps} from '../../app/redux'
-import {headerSelector as selector} from '../../selectors'
 import getStore from '../../app/store'
-import actionNames from './actionNames'
-import Header from './Header'
+import {headerSelector as selector} from '../../selectors'
+import Header from '../../components/header/Header'
+
+const actionNames = [
+  'showChannelMembersInvite',
+  'showInSidebar',
+  'hideSidebar',
+  'requestAddChannelToFavorites',
+  'requestRemoveChannelFromFavorites',
+  'updateMessageSearchQuery',
+  'hideIntercom'
+]
 
 const ConnectedHeader = connect(
   selector,
