+++ src/containers/header/index.js
@@ -1,0 +1,1 @@
+export HeaderProvider from './HeaderProvider'
