+++ src/containers/history/index.js
@@ -1,0 +1,1 @@
+export HistoryProvider from './HistoryProvider'
