+++ src/containers/invite-to-org/index.js
@@ -1,0 +1,1 @@
+export InviteToOrgProvider from './InviteToOrgProvider'
