+++ src/containers/navigation/index.js
@@ -1,0 +1,1 @@
+export NavigationProvider from './NavigationProvider'
