+++ src/containers/new-conversation/index.js
@@ -1,0 +1,1 @@
+export NewConversationProvider from './NewConversationProvider'
