+++ src/containers/notification-settings/index.js
@@ -1,0 +1,1 @@
+export NotificationSettingsProvider from './NotificationSettingsProvider'
