+++ src/containers/org-info/index.js
@@ -1,0 +1,1 @@
+export OrgInfoProvider from './OrgInfoProvider'
