+++ src/containers/sidebar/index.js
@@ -1,0 +1,1 @@
+export SidebarProvider from './SidebarProvider'
