+++ src/containers/typing-notification/index.js
@@ -1,0 +1,1 @@
+export TypingNotificationProvider from './TypingNotificationProvider'
