+++ src/containers/unread-channels/UnreadChannelsProvider.js
@@ -3,7 +3,7 @@ import {Provider, connect} from 'react-redux'
 
 import {unreadChannelsSelector as selector} from '../../selectors'
 import getStore from '../../app/store'
-import UnreadChannels from './UnreadChannels'
+import UnreadChannels from '../../components/unread-channels/UnreadChannels'
 
 const ConnectedUnreadChannels = connect(
   selector
