+++ src/containers/unread-channels/index.js
@@ -1,0 +1,1 @@
+export UnreadChannelsProvider from './UnreadChannelsProvider'
